+++ ec2_cluster/config/config.py
@@ -1,0 +1,4 @@
+
+
+class ClusterConfig:
+    pass
